% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor.R
\name{lambdaBF}
\alias{lambdaBF}
\title{Approximate Bayes factor for Pagel model}
\usage{
lambdaBF(trees, x, N_samples = 100, return_trace = F, importance_sampling = T)
}
\arguments{
\item{trees}{phylo object, or list of phylo objects}

\item{x}{named vector of observations for each tip}

\item{N_samples}{number of importance samples to draw}

\item{return_trace}{boolean, whether to return sampling traces.}

\item{importance_sampling}{boolean, whether to use importance sampling.}

\item{logarithm}{boolean, whether to return log likelihood.}
}
\value{
list containing Bayes factor estimate 'bf', as well as bayes factors
for each tree in 'bf_samples', optionally contains sampling traces
}
\description{
Uses importance sampling to approximate the Bayes factor of Pagel's lambda
(Pagel, 1999), aggregating across trees. Importance sampling used to
marginalise over lambda and makes use of a Laplace approximation to the
likelihood.
For some trees, the likelihood may not have a maximum in (0,1), in which case 
the Laplace approximation fails. In this case lambda is sampled uniformly.
}
\examples{
trees <- ape::rmtree(5, 100)
x <- phytools::fastBM(trees[[1]])
lambdaBF(trees, x)


}
